requirements.txt
dnspython==2.2.1
motor==3.1.1
psutil==5.9.4
pyromod==1.5
pyrofork
#pyrogram==2.0.106
python-dotenv==0.21.0
telethon==1.26.1
TgCrypto==1.2.5
pymongo==4.6.0
GitPython==3.1.40